Create Counts the number of unique  values
he Academics Updating Lists measure counts the number of unique Event Creator entries in the 'As Exported' table. If there are no Event Creator values (i.e., if the table is empty or the column contains no data), it returns 0. Otherwise, it returns the actual count of unique Event Creator values.

This logic ensures that you get a 0 instead of a blank value when there are no records to count, which can be useful in reports or dashboards where you want to avoid displaying blank values.